+++ modules/core/karrio/server/core/tests/base.py
@@ -1,4 +1,5 @@
 from django.contrib.auth import get_user_model
+from django.urls import reverse
 from rest_framework.test import APITestCase as BaseAPITestCase, APIClient
 from karrio.server.core.logging import logger